+++ src/ImobFeed.Core/Leitores/LeitorRecomendacaoNu.cs
@@ -21,7 +21,7 @@ public sealed class LeitorRecomendacaoNu : ILeitorRecomendacao
             if (codigo is null)
                 continue;
 
-            float? peso = LeitorCampo.LerPeso(line);
+            decimal? peso = LeitorCampo.LerPeso(line);
             if (peso is null)
                 continue;
